backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
numpy==1.24.3
pillow==10.4.0
requests==2.32.3
torch==2.3.1
torchvision==0.18.1
faiss-cpu==1.8.0.post1
python-multipart==0.0.9
tensorflow==2.13.0
tqdm==4.66.1
